requirements-dbc.txt
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
citibike_project @ file:///C:/Users/UCHE/Desktop/citibike_project/dist/citibike_project-0.0.1-py3-none-any.whl
colorama==0.4.6
comm==0.2.3
databricks-connect==15.4.0
databricks-sdk==0.63.0
debugpy==1.8.16
decorator==5.2.1
executing==2.2.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
packaging==25.0
pandas==2.3.1
parso==0.8.4
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzmq==27.0.1
requests==2.32.4
rsa==4.9.1
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
